feat: get_countries and get_themes added created
