+++ idbsocialdatapy/utils.py
@@ -1,9 +1,5 @@
 import os
 
-import pandas as pd
-
-# -----CORE FUNCTION--------
-
 
 def iadburls() -> dict:
     """
@@ -21,26 +17,3 @@ def iadburls() -> dict:
         geojson_url=geojson_url,
     )
     return url_dict
-
-
-# -----GET COUNTRIES FUNCTION----------
-
-
-def get_countries():
-    """
-    get_countries() creates a pandas dataframe.
-    """
-    url_links = iadburls()
-    get_countries_url = os.path.join(url_links["metadata_url"], "countries")
-    df = pd.read_json(get_countries_url)
-    return df
-
-
-# -----GET THEMES FUNCTION-----
-
-
-def get_themes():
-    url_links = iadburls()
-    get_themes_url = os.path.join(url_links["metadata_url"], "themes")
-    df = pd.read_json(get_themes_url)
-    return df
